requirements.txt
streamlit
pandas
pickle5
gdown
requests
python-dotenv